Fixed size issues with offsets and counts Fixed all bugs that would cause the standard tests to fail
Important: New files will be 8 bytes larger than old files

+++ LibTiff/Internal/TiffDirectory.cs
@@ -66,7 +66,7 @@ namespace BitMiracle.LibTiff.Classic.Internal
         /// </summary>
         public bool td_stripbytecountsorted;
 
-        public short td_nsubifd;
+        public int td_nsubifd;
         public long[] td_subifd;
 
         // YCbCr parameters
